+++ rust/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### 0.24.1
 
-Worker: Fix, use `thread_local` buffer on `MS_ABORT()` ([PR#1873](https://github.com/versatica/mediasoup/pull/1873)).
+- Worker: Fix, use `thread_local` buffer on `MS_ABORT()` ([PR#1873](https://github.com/versatica/mediasoup/pull/1873)).
+- Handle subchannels in pipe `DataConsumers` ([PR#XXXX](https://github.com/versatica/mediasoup/pull/XXXX)).
 
 ### 0.24.0